Enforce variables and outputs in specific files
From the product guide:
```
Declare all variables in variables.tf, including a description and type
Declare all outputs in outputs.tf, including a description
```

This adds a 'file layout' check that ensures resources appear in the correct files. We can extend or refactor some of this behavior later, if needed.